flask/flask_requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.19